airflow/requirements.txt
apache-airflow==2.7.0
pymysql==1.1.0
python-dotenv==1.0.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-core==1.30.0
azure-ai-textanalytics==5.3.0
pyarrow==14.0.1
pandas==2.0.3
fastavro==1.9.0
openpyxl==3.1.2
xmltodict==0.13.0

